+++ src/main/java/emu/nes/ops/TransferAccumulatorX.java
@@ -13,7 +13,7 @@ public class TransferAccumulatorX implements Operation {
 
     @Override
     public int execute(Registers registers, Bus bus, AddressingResult res) {
-        int data = registers.getAcc();
+        byte data = registers.getAcc();
         updateFlags(registers, data);
         registers.setX(data);
         return 0;
