+++ src/main/java/emu/nes/ops/TransferAccumulatorY.java
@@ -13,7 +13,7 @@ public class TransferAccumulatorY implements Operation {
 
     @Override
     public int execute(Registers registers, Bus bus, AddressingResult res) {
-        int data = registers.getAcc();
+        byte data = registers.getAcc();
         updateFlags(registers, data);
         registers.setY(data);
         return 0;
